# internal/model/model_long_link_request.go
/*
 * Swagger URL Shortener
 *
 * URL Shortener API
 *
 * API version: 1.0.0
 * Contact: admin@urlshortener.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

// LongLingRequest ...
type LongLinkRequest struct {
	LongLink string `json:"longLink,omitempty"`
}