# internal/model/model_short_link_response.go
/*
 * Swagger URL Shortener
 *
 * URL Shortener API
 *
 * API version: 1.0.0
 * Contact: admin@urlshortener.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

// ShortLinkResponse ...
type ShortLinkResponse struct {
	ShortLink string `json:"shortLink,omitempty"`
}